requirements.txt
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
Werkzeug==1.0.1
